2023-01
First part was very easy. Second part annoying because none of the
examples shed light on the corner cases such as "eightwo" which should
be 82. Ended up finding the hint/clue needed on Reddit. After knowing
the problem I could find these cases in the input file and make sure
it worked properly.