2022-09 and 2022-20 broken due to mismatch in use of MAP-2D-ARRAY
MAP-2D-ARRAY passes ROW and COL as the coordinates... but the rest of
the code expect X and Y (where X is a column and Y is a row). Thus we
need a small shim in both cases to swap the argument order from (ARRAY
ROW COL) to (ARRAY COL ROW).