2023.04
Part 2 started with a naive recursive solution of having a to-process
list and working down it, sometimes appending more work to it. This
worked OK with the example but was not going to be fast enough (less
than a couple minutes) on the input.

Searched for hints which I didn't understand, finally after reading
through a few solutions I started to get the idea of it. For each card
N with wins W add its count to cards N+1..N+W. Then process card N+1
until done. Now sum the counts.