2024-14 complete
For part 2 i started to just check each one then saw there was a
pattern o when a horizontal grouping would happen; when a vertical
grouping would happen. Speculated there was some point where they'd
line up to create an xmas tree.

However the math failed me and I did not search high enough. Digressed
into some hints about the math which still didn't work for me (likely
I was doing it wrong).

Finally followed up on the hint about minimizing the safety
factor (which was the computation in part1). This worked! when most
robots are grouped to create a tree (this was stated in the text
- *most robots*) then the safety factor would be lower.

So search for the minimum safety factor and visually verified that was
the correct one before submitting.